factors out library add command to its own module
