+++ src/cli/library.js
@@ -1,11 +1,7 @@
-import {LibraryAddCommand, LibraryAddCommandSite} from '../cmd';
 import ParticleApi from '../cmd/api';
 import settings from '../../settings';
 
-import chalk from 'chalk';
-import log from '../app/log';
-import {spin} from '../app/ui';
-
+import libraryAdd from './library_add';
 import libraryInstall from './library_install';
 import libraryMigrate from './library_migrate';
 import libraryInit from './library_init';
@@ -14,7 +10,6 @@ import librarySearch from './library_search';
 import libraryContribute from './library_upload';
 import libraryPublish from './library_publish';
 import libraryDelete from './library_delete';
-import {buildAPIClient} from './library_search';
 
 function api() {
 	if (!api._instance) {
@@ -25,43 +20,11 @@ function api() {
 	return api._instance;
 }
 
-export class CLILibraryAddCommandSite extends LibraryAddCommandSite {
-	constructor(argv, apiClient) {
-		super();
-		this._apiClient = apiClient;
-		[this.name, this.version='latest'] = argv.params.name.split('@');
-		this.dir = argv.params.dir || process.cwd();
-	}
-
-	apiClient() {
-		return this._apiClient;
-	}
-
-	libraryIdent() {
-		// todo - shouldn't this be a promise?
-		return {
-			name: this.name,
-			version: this.version
-		};
-	}
-
-	projectDir() {
-		return this.dir;
-	}
-
-	fetchingLibrary(promise, name) {
-		return spin(promise, `Adding library ${chalk.green(name)}`);
-	}
-
-	addedLibrary(name, version) {
-		return Promise.resolve().then(() => log.success(`Added library ${chalk.green(name)} ${version} to project`));
-	}
-}
-
 export default ({root, factory}) => {
 	const lib = factory.createCategory(root, 'library', 'Manages firmware libraries', { alias: 'libraries' });
 
-	libraryInit({root, lib, factory});
+	libraryAdd({lib, factory, apiJS: api()});
+	libraryInit({lib, factory});
 	libraryInstall({lib, factory, apiJS: api()});
 	libraryList({lib, factory, apiJS: api()});
 	libraryMigrate({lib, factory});
@@ -70,17 +33,5 @@ export default ({root, factory}) => {
 	libraryPublish({lib, factory, apiJS: api()});
 	libraryDelete({lib, factory, apiJS: api()});
 
-	// todo - move library add to its own module
-	factory.createCommand(lib, 'add', 'Add a library to the current project.', {
-		options: {},
-		params: '<name>',
-
-		handler: function libraryAddHandler(argv) {
-			// todo ... and here
-			const site = new CLILibraryAddCommandSite(argv, buildAPIClient(api()));
-			const cmd = new LibraryAddCommand();
-			return site.run(cmd);
-		}
-	});
 	return lib;
 };
